refactor: `UpdateFileAsset`
